add mixing length to precomputed quantities
