+++ Project.toml
@@ -68,7 +68,7 @@ IntervalSets = "0.5, 0.6, 0.7"
 Krylov = "0.9"
 LinearAlgebra = "1"
 Logging = "1"
-NCDatasets = "0.11, 0.12, 0.13"
+NCDatasets = "=0.13.1"
 NVTX = "0.3"
 Pkg = "1.8"
 Printf = "1"
