+++ config/model_configs/sphere_held_suarez_rhoe_equilmoist_topography_dcmip.yml
@@ -24,5 +24,5 @@ diagnostics:
     period: 3hours
   - short_name: hur
     period: 3hours
-  - short_name: mixlgm
+  - short_name: lmix
     period: 3hours
