+++ perf/flame.jl
@@ -37,8 +37,8 @@ ProfileCanvas.html_file(joinpath(output_dir, "flame.html"), results)
 #####
 
 allocs_limit = Dict()
-allocs_limit["flame_perf_target"] = 148_256
-allocs_limit["flame_perf_target_tracers"] = 180_512
+allocs_limit["flame_perf_target"] = 148_592
+allocs_limit["flame_perf_target_tracers"] = 180_848
 allocs_limit["flame_perf_target_edmfx"] = 7_005_552
 allocs_limit["flame_perf_diagnostics"] = 25_356_928
 allocs_limit["flame_perf_target_diagnostic_edmfx"] = 1_311_040
@@ -46,7 +46,7 @@ allocs_limit["flame_sphere_baroclinic_wave_rhoe_equilmoist_expvdiff"] =
     4_018_252_656
 allocs_limit["flame_perf_target_threaded"] = 1_276_864
 allocs_limit["flame_perf_target_callbacks"] = 37_277_112
-allocs_limit["flame_perf_gw"] = 3_226_429_472
+allocs_limit["flame_perf_gw"] = 3_226_430_064
 allocs_limit["flame_perf_target_prognostic_edmfx_aquaplanet"] = 1_258_848
 
 # Ideally, we would like to track all the allocations, but this becomes too
