+++ src/cache/precomputed_quantities.jl
@@ -50,6 +50,7 @@ function precomputed_quantities(Y, atmos)
         ᶜts = similar(Y.c, TST),
         ᶜp = similar(Y.c, FT),
         ᶜh_tot = similar(Y.c, FT),
+        ᶜmixing_length = similar(Y.c, FT),
         sfc_conditions = Fields.Field(SCT, Spaces.level(axes(Y.f), half)),
     )
     cloud_diagnostics = (; ᶜcloud_fraction = similar(Y.c, FT),)
@@ -68,7 +69,6 @@ function precomputed_quantities(Y, atmos)
             ᶜρ⁰ = similar(Y.c, FT),
             ᶜlinear_buoygrad = similar(Y.c, FT),
             ᶜstrain_rate_norm = similar(Y.c, FT),
-            ᶜmixing_length = similar(Y.c, FT),
             ᶜK_u = similar(Y.c, FT),
             ᶜK_h = similar(Y.c, FT),
             ρatke_flux = similar(Fields.level(Y.f, half), C3{FT}),
@@ -105,7 +105,6 @@ function precomputed_quantities(Y, atmos)
             ᶜtke⁰ = similar(Y.c, FT),
             ᶜlinear_buoygrad = similar(Y.c, FT),
             ᶜstrain_rate_norm = similar(Y.c, FT),
-            ᶜmixing_length = similar(Y.c, FT),
             ᶜK_u = similar(Y.c, FT),
             ᶜK_h = similar(Y.c, FT),
             ρatke_flux = similar(Fields.level(Y.f, half), C3{FT}),
@@ -345,6 +344,11 @@ NVTX.@annotate function set_precomputed_quantities!(Y, p, t)
 
     if !isnothing(p.sfc_setup)
         SurfaceConditions.update_surface_conditions!(Y, p, t)
+    end
+
+    if isnothing(turbconv_model)
+        (; ᶜmixing_length) = p.precomputed
+        compute_gm_mixing_length!(ᶜmixing_length, Y, p)
     end
 
     if turbconv_model isa PrognosticEDMFX
